requirements.txt
numpy
soundfile
scipy
torch
torchaudio
openai-whisper
